#4 added tests for deleting image after 10 commands, forwardsplashes, relatives paths, decoding of html strings

+++ lab5/lab5/task1/DocumentEditor/Documents/ImageHandler.cs
@@ -36,10 +36,9 @@ namespace task1.DocumentEditor.Documents
 
 		public void DeleteImage(string path)
 		{
-			if (File.Exists(path))
+			if (File.Exists(path) && _imagesForDeletion.Remove(path))
 			{
 				File.Delete(path);
-				_imagesForDeletion.Remove(path);
 			}
 		}
